requirements.txt
greenlet==3.0.3
schedule==1.2.2
speedtest-cli==2.1.3
SQLAlchemy==2.0.32
typing_extensions==4.12.2
